Feat: right header增加status bar
