+++ src/context/ProjectsContext.js
@@ -1,0 +1,3 @@
+import React from "react";
+
+export const ProjectsContext = React.createContext(null);
