--- src/context/syncMiddleZoneProjectContext.js
@@ -1,3 +1,0 @@
-import React from "react";
-
-export const SyncMiddleZoneProjectContext = React.createContext(null);
